Refactor code comments for consistency and clarity across multiple files

+++ tests/R4PDF.Tests/ColorParserTests.cs
@@ -1,5 +1,5 @@
-using R4PDF.Parsing;
 using PdfSharpCore.Drawing;
+using R4PDF.Parsing;
 
 namespace R4PDF.Tests;
 
@@ -9,18 +9,18 @@ public class ColorParserTests
     public void Parse_HexColor6_ReturnsCorrectColor()
     {
         var color = ColorParser.Parse("#FF0000");
-        Assert.Equal(255, (int)color.R);
-        Assert.Equal(0, (int)color.G);
-        Assert.Equal(0, (int)color.B);
+        Assert.Equal(255, color.R);
+        Assert.Equal(0, color.G);
+        Assert.Equal(0, color.B);
     }
 
     [Fact]
     public void Parse_HexColor3_ReturnsCorrectColor()
     {
         var color = ColorParser.Parse("#F00");
-        Assert.Equal(255, (int)color.R);
-        Assert.Equal(0, (int)color.G);
-        Assert.Equal(0, (int)color.B);
+        Assert.Equal(255, color.R);
+        Assert.Equal(0, color.G);
+        Assert.Equal(0, color.B);
     }
 
     [Theory]
@@ -30,27 +30,27 @@ public class ColorParserTests
     public void Parse_NamedColor_ReturnsCorrectColor(string name)
     {
         var color = ColorParser.Parse(name);
-        Assert.Equal(0, (int)color.R);
-        Assert.Equal(0, (int)color.G);
-        Assert.Equal(0, (int)color.B);
+        Assert.Equal(0, color.R);
+        Assert.Equal(0, color.G);
+        Assert.Equal(0, color.B);
     }
 
     [Fact]
     public void Parse_NullOrEmpty_ReturnsBlack()
     {
         var color = ColorParser.Parse(null);
-        Assert.Equal(0, (int)color.R);
-        Assert.Equal(0, (int)color.G);
-        Assert.Equal(0, (int)color.B);
+        Assert.Equal(0, color.R);
+        Assert.Equal(0, color.G);
+        Assert.Equal(0, color.B);
     }
 
     [Fact]
     public void Parse_WithDefault_ReturnsDefaultForNull()
     {
         var color = ColorParser.Parse(null, XColors.Red);
-        Assert.Equal(255, (int)color.R);
-        Assert.Equal(0, (int)color.G);
-        Assert.Equal(0, (int)color.B);
+        Assert.Equal(255, color.R);
+        Assert.Equal(0, color.G);
+        Assert.Equal(0, color.B);
     }
 
     [Fact]
